chore(tenderize): extract types into common
