+++ src/apps/tenderize/common/tenderize.swap.token-fetcher.ts
@@ -13,13 +13,7 @@ import {
 import { TenderizeContractFactory, TenderToken } from '../contracts';
 
 import { TenderizeTokenDefinitionsResolver } from './tenderize.token-definition-resolver';
-
-export type TenderizeTokenDefinition = {
-  address: string;
-  steak: string;
-  lpToken: string;
-  tenderSwap: string;
-};
+import { TenderizeTokenDefinition } from './types';
 
 export abstract class SwapTokenFetcher extends AppTokenTemplatePositionFetcher<
   TenderToken,
