+++ src/apps/tenderize/common/types.ts
@@ -1,0 +1,7 @@
+export type TenderizeTokenDefinition = {
+  id: string;
+  address: string;
+  steak: string;
+  lpToken: string;
+  tenderSwap: string;
+};
